requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
seqeval>=1.2.2

# LLM API clients
openai>=1.12.0
anthropic>=0.39.0

# Environment and config
python-dotenv>=1.0.0

# Data and reporting
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tabulate>=0.9.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.66.0
tenacity>=8.2.0
